refacto: split datasetTypes into multiple files
