+++ ui/components/core/src/lib/types/dataset/annotations/index.ts
@@ -1,0 +1,6 @@
+export * from "./Annotation";
+export * from "./Bbox";
+export * from "./Keypoint";
+export * from "./Mask";
+export * from "./NamedEntity";
+export * from "./Tracklet";
