+++ ui/components/core/src/lib/types/dataset/entities/index.ts
@@ -1,0 +1,2 @@
+export * from "./Entity";
+export * from "./Track";
