+++ ui/components/core/src/lib/types/dataset/index.ts
@@ -1,1 +1,8 @@
-export * from "./NamedEntity";
+export * from "./annotations";
+export * from "./DatasetInfo";
+export * from "./DatasetItem";
+export * from "./datasetTypes";
+export * from "./entities";
+export * from "./items";
+export * from "./sources";
+export * from "./views";
