+++ ui/components/core/src/lib/types/dataset/sources/Source.ts
@@ -1,0 +1,18 @@
+import { z } from "zod";
+import { BaseData, type BaseDataFields } from "../datasetTypes";
+
+const sourceSchema = z
+  .object({
+    name: z.string(),
+    kind: z.string(),
+    metadata: z.string(),
+  })
+  .strict();
+export type sourceType = z.infer<typeof sourceSchema>;
+
+export class Source extends BaseData<sourceType> {
+  constructor(obj: BaseDataFields<sourceType>) {
+    sourceSchema.parse(obj.data);
+    super(obj);
+  }
+}
