+++ ui/components/core/src/lib/types/index.ts
@@ -5,7 +5,6 @@ License: CECILL-C
 -------------------------------------*/
 
 export * from "./dataset";
-export * from "./datasetTypes";
 export * from "./modelsTypes";
 export * from "./objectTypes";
 export * from "./toolsTypes";
